fix(verbs): break the executor's import ring and make its steering args honest
The bounded executor reached back for two things it did not own, and the two
reaches closed a runtime ring with the dispatch core and the orchestrate
composite. Both are now injected or extracted rather than imported:

- The handler table is passed IN by the composite entry that owns it, so the
  dynamic `import('../composite.js')` is gone and `handlers` is required.
- ActionId admission moves to `dispatch/core/dispatch-admission.ts`, imported
  by both the dispatch path and the executor. A move, not a rewrite; the
  dispatch module keeps exporting what its consumers import, and reaches the
  evaluator through the same leaf.

Two type assertions go with them: the operation payload is typed at its
binding instead of asserted across the append seam, and the request's `args`
narrow through a type guard rather than a cast.

Subject identity is now written LAST when a leaf's arguments are built. Merged
before the step params, a runbook param spelled `streamId`/`featureId` would
overwrite the subject, and the leaf would commit to one stream while emission
verification watched another.

`riskTier` and `boundaryTouching` become required in the task-completion
argument schema. Every gate step in that runbook passes them and the compiler
refuses an unbindable placeholder, so declaring them optional told a caller to
omit what the call would then refuse. The compiler's own refusal keeps its
teeth over a fixture intent whose schema still leaves the variable optional —
the population it exists to protect.

Comments carrying roadmap ordinals are restated in words, the three in-scope
execute suites declare their oracle authorities, and the tools-list golden plus
the CLI-surface and proof-fixture baselines are regenerated for the contract
and description changes.
